завдання 3
