+++ answer.txt
@@ -1,9 +1,9 @@
 ----------------------------------------------------------------------------------------------------
 id: 1
-username: dfa
+username: ad
 birth_date: 1.1.1
-start_date: 16.1.1
-end_date: 17.1.1
-vaccine: pfizer
+start_date: 12.2.3
+end_date: 13.2.3
+vaccine: moderna
 international_passport: aa000000
 ----------------------------------------------------------------------------------------------------
